refactor from milestone 2
