+++ src/assets/styles/index.css
@@ -1,7 +1,7 @@
 /* universal */
 html,
 body {
-  font-size: 12px;
+  font-size: 16px;
   overflow-x: hidden;
 }
