+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import App from './containers/App';
+import App from './components/App';
 import configureStore from './configureStore';
 import initialState from './rawData';
